server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "google-auth>=2.49.1",
    "httpx>=0.28.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.27.2.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "requests>=2.33.1",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
